tests/submission/PhilipTaylor/canvas/security.pattern.canvas.timing.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: security.pattern.canvas.timing</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.security.html">security</a>.<a href="index.security.pattern.html">pattern</a>.<a href="index.security.pattern.canvas.html">canvas</a>.timing</h1>
<p class="desc">Pattern safety depends on whether the source was origin-clean, not on whether it still is clean</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.security.start">security.start</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.security.fillStyle.canvas">security.fillStyle.canvas</a>
</li></ul>
</div>
<p class="notes">Disagrees with spec on "is" vs "was"
</p><p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Pattern safety depends on whether the source was origin-clean, not on whether it still is clean");
_addTest(function(canvas, ctx) {

var canvas2 = document.createElement('canvas');
canvas2.width = 100;
canvas2.height = 50;
var ctx2 = canvas2.getContext('2d');
ctx2.fillStyle = '#0f0';
ctx2.fillRect(0, 0, 100, 50);
var p = ctx.createPattern(canvas2, 'repeat');
ctx2.drawImage(document.getElementById('yellow.png'), 0, 0); // make canvas2 origin-unclean
ctx.fillStyle = p;
ctx.fillRect(0, 0, 100, 50);
canvas.toDataURL();
ctx.getImageData(0, 0, 1, 1);
_assert(true, "true"); // okay if there was no exception


});
</script>
<img class="resource" id="yellow.png" src="http://www2.w3c-test.org/html/tests/approved/images/yellow.png">

<div id="log"></div>
</body></html>
